core: remove tracing mod
