+++ core/src/lib.rs
@@ -60,8 +60,6 @@ cfg_client_or_server! {
 	pub mod middleware;
 }
 
-/// Shared tracing helpers to trace RPC calls.
-pub mod tracing;
 pub use async_trait::async_trait;
 pub use error::{RegisterMethodError, StringError};
